IS Project/backend/requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
# Other dependencies can be added here